==> 0_reviews_analysis.ipynb <==
# DATA ANALYSIS - AMAZON REVIEWS
____
by Vitor Flisch Cavalanti<br>
May 2021

<b>Case study Sr. Business Analyst</b>

<hr><h2>1. Working with Reviews file</h2><hr>

import pandas as pd
import numpy as np 
import ast

# use this file_name variable to receive as parameter for future automation
file_name = "reviews_Movies_and_TV_5.csv"
reviews = pd.read_csv("../files/" + file_name, nrows=5)

reviews.head()

# reading just columns that will be used
reviews = pd.read_csv("../files/" + file_name, usecols = ['reviewerID','asin','helpful','overall','reviewTime']) 

reviews["file"]=file_name

# checking nulls
reviews.isnull().sum()

# converting reviewTime to date format
reviews['reviewTime'] =  pd.to_datetime(reviews['reviewTime'], format='%m %d, %Y')

# transforming helpful to a list object
reviews["helpful"] = reviews["helpful"].apply(eval)

# calculating helpful score (numerator / denominator )
reviews['helpful_score'] = reviews.helpful.apply(lambda x: x[0]) / reviews.helpful.apply(lambda x: x[1])

reviews.head(20)

# count by review
reviews['overall'].value_counts() 

# avg review
reviews['overall'].mean() 

# save adjusted file to exports folder
reviews.to_csv("../exports/" + file_name)
