==> 1_metadata_analysis.ipynb <==
# DATA ANALYSIS - AMAZON REVIEWS
____
by Vitor Flisch Cavalanti<br>
May 2021

<b>Case study Sr. Business Analyst</b>

# importing libraries
import pandas as pd
import numpy as np 
import ast

<hr><h2>1. Working with Metadata file</h2><hr>

metadata = pd.read_csv("../files/metadata.csv", nrows=5)

metadata.head()

metadata = pd.read_csv("../files/metadata.csv", usecols = ['asin','salesRank','categories','title','related','brand','price'])

metadata.info()

# checking nulls
metadata.isnull().sum() 

metadata.shape

<hr><h4>a. Formatting Sales Rank</h4><hr>

# Filling Null values at the Sales Rank column
metadata.salesRank.fillna("{'Null': 0}", inplace=True)

# replacing empty sales rank with {'Null': 0}
metadata['salesRank'] = metadata['salesRank'].replace(["{}"],"{'Null': 0}")

# converting sales rank to dict
metadata["salesRank"] = metadata.salesRank.apply(lambda x: ast.literal_eval(x))

# Getting category from sales rank category 
metadata["salesrank_category"] = metadata.salesRank.apply(lambda x: list(x.keys())[0]) 

metadata["salesrank_value"] = metadata.salesRank.apply(lambda x: list(x.values())[0]) 

# top 5 categories with more products
metadata['salesrank_category'].value_counts().head(5) 

# Checking Duplicated Products
duplicateRows = metadata[metadata.duplicated(['asin'])] 
duplicateRows['asin'].count() # checking dupplicated products

del duplicateRows

#drop sales rank column
metadata.drop(['salesRank'], axis='columns',inplace=True)

<hr><h4>b. Exploding Categories</h4><hr>

metadata_cat = metadata[~metadata['categories'].isna()]
metadata_cat

# converting categories to list
metadata_cat['categories'] = metadata_cat.categories.apply(eval)

# create auxiliar data frame to explode categories
metadata_categories = (metadata_cat[["asin","categories"]])

# exploding categories by product
metadata_categories = metadata_categories.explode("categories").explode("categories")

# getting the most common categories by product
group_metadata_cat = metadata_categories.groupby(['asin', 'categories']).size().reset_index(name="Count")

# filter the categorie(s) with most occurrences 
filter_max_cat = group_metadata_cat[group_metadata_cat.groupby(["asin"])['Count'].transform(max) == group_metadata_cat["Count"]]

# get the first category in case of having more than 1
filter_first_cat = filter_max_cat.groupby("asin").first().reset_index()

#cleaning unused data frames
del metadata_categories
del group_metadata_cat
del filter_max_cat
del metadata_cat

# merging with the main dataframe
metadata_join = pd.merge(metadata,filter_first_cat,on="asin")

# cleaning unused columns and the old dataframe
metadata_join.drop(['categories_x','Count'], axis='columns',inplace=True)
metadata_join.rename(columns = {'categories_y': 'main_category'}, inplace = True)
del metadata

<hr><h4>c. Extracting and Exploding Bought-Together</h4><hr>

metadata_join['related'].apply(type)

# remove products with nulls at related column
metadata_related = metadata_join[~metadata_join["related"].isna()]

 # remove products without bought_together
 metadata_related = metadata_related[metadata_related['related'].str.contains("bought_together")]

metadata_related.drop(['title','price','brand','salesrank_category','salesrank_value','main_category'], axis='columns',inplace=True)

# convert related column to dict
metadata_related["related"] = metadata_related.related.apply(lambda x: ast.literal_eval(x))

# creating new column only for bought together
metadata_related["bought_together"] = metadata_related.related.apply(lambda x: list(x.get("bought_together"))) 

# deleting related column and merging back to main dataframe
metadata_related.drop(['related'], axis='columns',inplace=True)

# merging back to the main dataframe and deleting the related column
# metadata_join = pd.merge(metadata_join,metadata_related,on="asin")

metadata_join

# exploding bought together
metadata_related = metadata_related.explode("bought_together")

metadata_related

metadata_join.drop(["related"], axis="columns", inplace=True)

metadata_join.to_csv('../exports/metadata_full.csv')

metadata_join.shape

metadata_related.to_csv('../exports/metadata_related_full.csv')
