==> 2_merge_files.ipynb <==
# DATA ANALYSIS - AMAZON REVIEWS
____
by Vitor Flisch Cavalanti<br>
May 2021

<b>Case study Sr. Business Analyst</b>

<hr><h2>1. Concatenating Reviews</h2><hr>

import pandas as pd
import numpy as np 
import glob 

# loading all reviews*.csv
path = r'..\exports'                     
all_files = glob.glob(os.path.join(path, "reviews*.csv"))    
df_from_each_file = (pd.read_csv(f) for f in all_files)
concatenated_df   = pd.concat(df_from_each_file, ignore_index=True)

concatenated_df.info()

concatenated_df.count()

asin_df = concatenated_df[['asin','file']]

concatenated_df.drop(['file'], axis="columns",inplace=True)

# saving concatenated reviews to final folder
concatenated_df.to_csv("../exports/final/reviews_concatenated.csv")

# removing duplicates
asin_df = asin_df.drop_duplicates()

# checking number of unique products with reviews
asin_df.count()

# opening metadata
metadata = pd.read_csv("../exports/metadata_full.csv")

# joining metadata with unique list of products with reviews
merge = pd.merge(asin_df, metadata,on=['asin'],  how='inner')

merge.count()

# saving metadata to final folder
merge.to_csv("../exports/final/metadata_final.csv")
del merge
del metadata
del asin_df

<hr><h2>2. Get Overall review by year for Products Bought Together</h2><hr>

# create separate year column for future aggregation
concatenated_df['reviewYear'] = concatenated_df['reviewTime'].str.slice(start=0, stop=4)

# create aggregated data frame by product / year and average review
df_agg = concatenated_df.groupby(['asin','reviewYear'])['overall'].mean().reset_index()

df_agg.head()

# read metadata related file
metadata_related = pd.read_csv("../exports/metadata_related_full.csv")

# join metadata_related with annual average by product
metadata_join = pd.merge(metadata_related, df_agg, on=['asin'],  how='inner')

# renaming columns to simplify future joins
metadata_join.rename(columns = {'asin': 'asin_1', 'overall': 'overall_1'}, inplace = True)
metadata_join.rename(columns = {'bought_together': 'asin'}, inplace = True)

# join for the bought together products
metadata_join2 = pd.merge(metadata_join, df_agg, on=['asin','reviewYear'],  how='inner')

# renaming columns and save csv
metadata_join2.rename(columns = {'asin_1': 'asin_prod',  'asin': 'asin_bought_together', 'overall_1': 'overall_prod', 'overall': 'overall_bought_together'}, inplace = True)
metadata_join2.to_csv("../exports/final/metadata_related.csv")
del concatenated_df
del df_agg
del metadata_join
del metadata_join2
del metadata_related
