==> 3_data_analysis.ipynb <==
# DATA ANALYSIS - AMAZON REVIEWS
____
by Vitor Flisch Cavalanti<br>
May 2021

<b>Case study Sr. Business Analyst</b>

Please present your analysis and actionable insights, as if you were to present to the management of Amazon. The "Actionable business insights" should be based on data, an example of an analytical approach is as follows:
 
1) What is the relation between the reviews and the helpfulness?<br>
2) What is the review behavior among different categories?<br>
3) Is there a relationship between price and reviews?<br>
4) Which group of reviewers is more valuable to the business?<br>
5) Is there a relation between reviews from products which are bought together?<br>
6) Optional - Any other hypothesis you think is interesting, as long it would have business value for Amazon

import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt

# read reviews
reviews = pd.read_csv("../exports/final/reviews_concatenated.csv")

# read metadata
metadata = pd.read_csv("../exports/final/metadata_final.csv")

# read metadata_related
metadata_related = pd.read_csv("../exports/final/metadata_related.csv")

merge = pd.merge(reviews, metadata, on="asin", how="inner")

<hr><h3>1) What is the relation between the reviews and the helpfulness?</h3><hr>

corr = reviews[~reviews.helpful_score.isna()][['helpful_score','overall']].corr(method='pearson')
corr.style.background_gradient()

# correlation by group of products
merge[~merge.helpful_score.isna()].groupby('file')[['helpful_score','overall']].corr(method='pearson').unstack().iloc[:,1]

corr = reviews[~reviews.helpful_score.isna()][['helpful_score','overall']].corr(method='spearman')
corr.style.background_gradient()

# correlation by group of products
merge[~merge.helpful_score.isna()].groupby('file')[['helpful_score','overall']].corr(method='spearman').unstack().iloc[:,1]

# Removing unused columns
merge.drop(["Unnamed: 0_x","Unnamed: 0.1_x","Unnamed: 0_y","Unnamed: 0.1_y"], axis="columns", inplace=True)

# Correlation Matrix
corr = merge.corr()
corr.style.background_gradient()

<hr><h3>3) Is there a relationship between price and reviews?</h3><hr>

corr = merge[['price','overall']].corr(method='pearson')
corr.style.background_gradient()

corr = merge[merge.price<=100][['price','overall']].corr(method='pearson')
corr.style.background_gradient()

# correlation by group of products
merge.groupby('file')[['price','overall']].corr(method='pearson').unstack().iloc[:,1]

# correlation by group of products
merge[merge.price<=100].groupby('file')[['price','overall']].corr(method='pearson').unstack().iloc[:,1]

# correlation by group of products
merge[merge.price>100].groupby('file')[['price','overall']].corr(method='pearson').unstack().iloc[:,1]

corr = merge[['price','overall']].corr(method='spearman')
corr.style.background_gradient()

corr = merge[merge.price<=100][['price','overall']].corr(method='spearman')
corr.style.background_gradient()

# correlation by group of products
merge.groupby('file')[['price','overall']].corr(method='spearman').unstack().iloc[:,1]

# correlation by group of products
merge[merge.price<=100].groupby('file')[['price','overall']].corr(method='spearman').unstack().iloc[:,1]

# correlation by group of products
merge[merge.price>100].groupby('file')[['price','overall']].corr(method='spearman').unstack().iloc[:,1]

<hr><h3>5) Is there a relation between reviews from products which are bought together?</h3><hr>

metadata_related.rename(columns = {'asin_prod': 'asin'}, inplace = True)

metadata_related_join = pd.merge(metadata,metadata_related, on="asin", how="inner")

corr = metadata_related[['overall_prod','overall_bought_together']].corr(method='pearson')
corr.style.background_gradient()

# pearson correlation by product group
metadata_related_join.groupby('file')[['overall_prod','overall_bought_together']].corr(method='pearson').unstack().iloc[:,1]

corr = metadata_related[['overall_prod','overall_bought_together']].corr(method='spearman')
corr.style.background_gradient()

# spearman correlation by product group
metadata_related_join.groupby('file')[['overall_prod','overall_bought_together']].corr(method='spearman').unstack().iloc[:,1]
